requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pyyaml==6.0.1
python-multipart==0.0.6

